feat: rename project and path from portfolio to showcase, make sure it run in andro and ios
