+++ README.md
@@ -1,6 +1,6 @@
-# Portfolio
+# Showcase
 
-Showcase of Fakhry's Portfolio.
+Showcase of Fakhry Mubarak's Portfolio.
 
 ## Stack
 - Dev Tools : Flutter 3.6.0
@@ -16,7 +16,7 @@ I’ve been exploring multiplatform development while working on a portfolio web
 
 I started by experimenting with Kotlin Multiplatform Mobile (KMM), which is still in its alpha stage for web support. I built a simple project and deployed it to [https://fakhrymubarak.github.io/kmm-hello-world/pages](https://fakhrymubarak.github.io/kmm-hello-world/pages). However, when I tried accessing it on an iOS web browser, it didn’t work as expected. This setback pushed me to explore other options.
 
-That’s when I decided to give Flutter a shot for this use case. With Flutter’s stable support for web, I created a similar portfolio website and deployed it to [https://fakhrymubarak.github.io/flutter-multiplatform/pages](https://fakhrymubarak.github.io/flutter-multiplatform/pages). To my relief, it worked perfectly across all platforms, including iOS.
+That’s when I decided to give Flutter a shot for this use case. With Flutter’s stable support for web, I created a experiment website and deployed it to [https://fakhrymubarak.github.io/flutter-multiplatform/pages](https://fakhrymubarak.github.io/flutter-multiplatform/pages). It worked perfectly across all platforms, including iOS.
 
 This experience solidified my decision to use Flutter for creating my portfolio website. And so, here we are!
