+++ android/app/src/main/kotlin/com/fakhry/showcase/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.fakhry.portfolio
+package com.fakhry.showcase
 
 import io.flutter.embedding.android.FlutterActivity
