+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:portfolio/core/ui/scroll/smooth_scroll_behaviour.dart';
-import 'package:portfolio/themes/themes.dart';
-import 'package:portfolio/ui/pages/dashboard/portfolio_web_page_v1.dart';
+import 'package:showcase/core/ui/scroll/smooth_scroll_behaviour.dart';
+import 'package:showcase/themes/themes.dart';
+import 'package:showcase/ui/pages/dashboard/portfolio_web_page_v1.dart';
 
 void main() {
   runApp(const MyApp());
