+++ lib/ui/pages/dashboard/components/custom_background_section_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:portfolio/themes/themes.dart';
+import 'package:showcase/themes/themes.dart';
 
 class CustomBackgroundSectionWidget extends StatelessWidget {
   const CustomBackgroundSectionWidget({required this.child, super.key});
