+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: portfolio
+name: showcase
 description: "Showcase of Fakhry Mubarak's Portfolio."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
